Refactor as a python module
