+++ .gitignore
@@ -3,3 +3,5 @@ venv
 .snakemake
 __pycache__
 galop_assembly
+*.egg-info
+build
