--- workflow/rules/hello_world.smk
@@ -1,3 +1,0 @@
-rule hello_world:
-    output: "hello_world.txt"
-    shell: "echo 'Hello, World!' > {output}"
